@@ -21,19 +21,19 @@
   "private": true,
   "dependencies": {
     "@angular-devkit/core": "^0.0.29",
-    "@angular/animations": "5.2.3",
+    "@angular/animations": "6.1.10",
     "@angular/cdk": "^5.0.4",
-    "@angular/common": "5.2.3",
-    "@angular/compiler": "5.2.3",
-    "@angular/core": "5.2.3",
+    "@angular/common": "6.1.10",
+    "@angular/compiler": "6.1.10",
+    "@angular/core": "6.1.10",
     "@angular/flex-layout": "2.0.0-beta.10",
-    "@angular/forms": "5.2.3",
-    "@angular/http": "5.2.3",
+    "@angular/forms": "6.1.10",
+    "@angular/http": "6.1.10",
     "@angular/material": "^5.0.4",
     "@angular/material-moment-adapter": "^5.0.4",
-    "@angular/platform-browser": "5.2.3",
-    "@angular/platform-browser-dynamic": "5.2.3",
-    "@angular/router": "5.2.3",
+    "@angular/platform-browser": "6.1.10",
+    "@angular/platform-browser-dynamic": "6.1.10",
+    "@angular/router": "6.1.10",
     "@ngx-translate/core": "9.0.2",
     "@ngx-translate/http-loader": "^2.0",
     "angular2-notifications": "1.0.4",
@@ -47,14 +47,15 @@
     "ngx-md": "3.1.0",
     "object-assign": "^4.1.1",
     "raven-js": "3.21.0",
-    "rxjs": "5.5.6",
+    "rxjs": "6.6.3",
     "typeface-montserrat": "^0.0.45",
     "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.4"
+    "zone.js": "^0.8.4",
+    "rxjs-compat": "^6.0.0-rc.0"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6",
-    "@angular/compiler-cli": "5.2.3",
+    "@angular/cli": "^6.2.9",
+    "@angular/compiler-cli": "6.1.10",
     "@types/jasmine": "2.8.3",
     "@types/jwt-simple": "^0.5.33",
     "@types/node": "~9.3.0",
@@ -73,6 +74,6 @@
     "protractor": "~5.2.2",
     "ts-node": "^4.1.0",
     "tslint": "^5.9.1",
-    "typescript": "^2.6.0"
+    "typescript": "^2.8.0"
   }
 }
