@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 import {HttpErrorResponse, HttpEvent, HttpHandler, HttpInterceptor, HttpRequest} from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {isNotEmpty} from '../util/helper';
 import {ErrorObservable} from 'rxjs/observable/ErrorObservable';
 import {NotificationWrapperService} from '../notifications/notification-wrapper.service';
