@@ -1,6 +1,8 @@
+
+import {map} from 'rxjs/operators';
 import {Injectable} from '@angular/core';
 import {ManageService} from '../../../manage/service/manage.service';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {ManageDto} from '../../../manage/model/manage.dto';
 import {InstitutionService} from '../../../manage/service/institution.service';
 
@@ -11,9 +13,9 @@ export class MockManageService extends ManageService {
     }
 
     public getManageData(): Observable<ManageDto> {
-        return this._institutionService.readAll()
-            .map(i => {
+        return this._institutionService.readAll().pipe(
+            map(i => {
                 return {institutions: i} as ManageDto
-            });
+            }));
     }
 }
