@@ -12,9 +12,8 @@ import {Util} from '../core/util/util';
 import {SchoolClassSubjectDto} from '../task/model/school-class-subject.dto';
 import {ActivatedRoute, Router} from '@angular/router';
 import {DateUtil} from '../core/services/date-util.service';
-import {Observable} from 'rxjs/Observable';
+import {Observable, Subject} from 'rxjs';
 import {examByNameComparator} from '../core/util/comparator';
-import {Subject} from 'rxjs/Subject';
 import {MarkService} from 'app/mark/service/mark.service';
 import * as moment from 'moment';
 import {Moment} from 'moment';
