@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 import {SchoolYearDto} from '../model/manage.dto';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {AppCrudService} from '../../core/services/core/app-crud.service';
 import {HttpClient} from '@angular/common/http';
 
