@@ -5,9 +5,8 @@ import {isNotEmpty, isTruthy} from '../../core/util/helper';
 import {MarkService} from '../service/mark.service';
 import {ConfirmDialogService} from '../../core/services/confirm-dialog.service';
 import {TranslateService} from '@ngx-translate/core';
-import {Observable} from 'rxjs/Observable';
+import {Observable, Subject} from 'rxjs';
 import {MarkGroupDto} from '../model/mark-group.dto';
-import {Subject} from 'rxjs/Subject';
 import {ResponsiveHelperService} from '../../core/services/ui/responsive-helper.service';
 import {DateUtil} from '../../core/services/date-util.service';
 import {MarkDto} from 'app/mark/model/mark.dto';
