@@ -1,7 +1,10 @@
+
+import {of as observableOf, Observable} from 'rxjs';
+
+import {catchError} from 'rxjs/operators';
 import {Injectable} from '@angular/core';
 import {ActivatedRouteSnapshot, Resolve, Router, RouterStateSnapshot} from '@angular/router';
 import {ExamDto} from '../model/exam.dto';
-import {Observable} from 'rxjs/Observable';
 import {ExamService} from './exam.service';
 import {HttpStatus} from '../../core/http/http-status';
 import {NotificationWrapperService} from '../../core/notifications/notification-wrapper.service';
@@ -16,16 +19,16 @@ export class ExamResolver implements Resolve<ExamDto> {
 
     resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<ExamDto> | Promise<ExamDto> | ExamDto {
         let id: number = route.params['id'];
-        return this._examService.readById(id)
-            .catch(error => {
+        return this._examService.readById(id).pipe(
+            catchError(error => {
                 if (error.status == HttpStatus.NOT_FOUND) {
                     this._notificationService.remove();
                     this._router.navigate(['/exam']).then(() =>
                         this._notificationService.error('i18n.modules.exam.notification.error.examNotFound.title', 'i18n.modules.exam.notification.error.examNotFound.message')
                     );
                 }
-                return Observable.of(null);
-            });
+                return observableOf(null);
+            }));
     }
 
 }
