@@ -1,3 +1,5 @@
+
+import {map} from 'rxjs/operators';
 import {Component, OnInit, ViewChild} from '@angular/core';
 import {MarkGroupDto} from '../model/mark-group.dto';
 import {FormBuilder, FormGroup, Validators} from '@angular/forms';
@@ -5,14 +7,13 @@ import {MarkService, WEIGHT_PATTERN} from '../service/mark.service';
 import {MarkDto} from '../model/mark.dto';
 import {ActivatedRoute, Router} from '@angular/router';
 import {ParentLinkedCreateUpdateComponent} from '../../core/common/parent-linked-create-update-component';
-import {Subject} from 'rxjs/Subject';
+import {Subject, Observable} from 'rxjs';
 import {isEmpty, isNotEmpty, isNotNull, isTrue} from '../../core/util/helper';
 import {Util} from '../../core/util/util';
 import {FormUtil} from '../../core/util/form-util';
 import {ViewMode} from '../../core/common/view-mode';
 import {ConfirmDialogService} from '../../core/services/confirm-dialog.service';
 import {MatSelect} from '@angular/material';
-import {Observable} from 'rxjs/Observable';
 import {NotificationWrapperService} from '../../core/notifications/notification-wrapper.service';
 
 @Component({
@@ -109,8 +110,8 @@ export class MarkGroupCreateUpdateComponent extends ParentLinkedCreateUpdateComp
 
     public submit() {
         if (this._markGroupCreateUpdateForm.valid && this._markGroupCreateUpdateForm.dirty) {
-            this._markService.saveMarkGroup(this._formToMarkGroup(this._markGroupCreateUpdateForm))
-                .map(() => Observable.fromPromise(this._goToSemesterView()))
+            this._markService.saveMarkGroup(this._formToMarkGroup(this._markGroupCreateUpdateForm)).pipe(
+                map(() => Observable.fromPromise(this._goToSemesterView())))
                 .filter(isTrue)
                 .subscribe(() => this._notificationService.success('i18n.common.notification.success.save', 'i18n.modules.mark.group.createUpdate.notification.success.message'));
         }
