@@ -1,5 +1,8 @@
+
+import {of as observableOf, Observable} from 'rxjs';
+
+import {switchMap, map} from 'rxjs/operators';
 import {ActivatedRouteSnapshot, Resolve, RouterStateSnapshot} from '@angular/router';
-import {Observable} from 'rxjs/Observable';
 import {Injectable} from '@angular/core';
 import {MarkService} from './mark.service';
 import {MarkGroupDto} from '../model/mark-group.dto';
@@ -10,9 +13,9 @@ export class SubjectMarkGroupResolver implements Resolve<MarkGroupDto> {
     }
 
     resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): MarkGroupDto | Observable<MarkGroupDto> | Promise<MarkGroupDto> {
-        return Observable.of(route.paramMap)
-            .map(map => parseInt(map.get('subjectId')))
-            .switchMap(subjectId => this._markService.getMarkGroupBySubjectId(subjectId));
+        return observableOf(route.paramMap).pipe(
+            map(map => parseInt(map.get('subjectId'))),
+            switchMap(subjectId => this._markService.getMarkGroupBySubjectId(subjectId)),);
     }
 
 }
