@@ -1,6 +1,7 @@
+
+import {of as observableOf, Observable} from 'rxjs';
 import {Injectable} from '@angular/core';
 import {TaskDto} from '../model/task.dto';
-import {Observable} from 'rxjs/Observable';
 import {CacheableCrudService} from '../../core/services/core/cacheable-crud.service';
 import {TaskProgressUpdateDto} from '../model/task-update-progress.dto';
 import {HttpClient} from '@angular/common/http';
@@ -14,7 +15,7 @@ export class TaskService extends CacheableCrudService<TaskDto, TaskDto[]> {
     public readById(id: number): Observable<TaskDto> {
         if (this.hasCache()) {
             let task = this.cache.find(task => task.id == id);
-            if (task) return Observable.of(task);
+            if (task) return observableOf(task);
         }
         return super.readById(id);
     }
